+++ src/core/AnalyserNode.cpp
@@ -10,7 +10,7 @@
 
 namespace lab {
 
-AnalyserNode::AnalyserNode(float sampleRate, size_t fftSize) : AudioBasicInspectorNode(sampleRate, 2), m_analyser(fftSize)
+AnalyserNode::AnalyserNode(float sampleRate, size_t fftSize) : AudioBasicInspectorNode((uint32_t) sampleRate, 2), m_analyser((uint32_t) fftSize)
 {
     //N.B.: inputs and outputs added by AudioBasicInspectorNode... no need to create here.
     setNodeType(NodeTypeAnalyser);
